Refactor omnistat-inspect
