+++ omnistat/inspect/__init__.py
@@ -4,5 +4,5 @@ Provides structured JSON output for hypothesis-driven GPU telemetry analysis.
 Designed for use with the analyze-job SKILL in agentic workflows.
 """
 
-from omnistat.inspect._analyzer import AnalyzeJob, QueryLedger
-from omnistat.inspect._cli import build_parser, main
+from omnistat.inspect.inspector import JobInspector, QueryLedger
+from omnistat.inspect.cli import build_parser, main
