+++ omnistat/inspect/cli.py
@@ -4,8 +4,8 @@ import argparse
 import json
 import sys
 
-from omnistat.inspect._analyzer import AnalyzeJob
-from omnistat.inspect._commands import (
+from omnistat.inspect.inspector import JobInspector
+from omnistat.inspect.commands import (
     cmd_counters,
     cmd_data_check,
     cmd_db_info,
@@ -17,7 +17,7 @@ from omnistat.inspect._commands import (
     cmd_stats,
     cmd_timeseries,
 )
-from omnistat.inspect._output import _append_query_log, _output_json, _write_scratch
+from omnistat.inspect.output import _append_query_log, _output_json, _write_scratch
 
 
 def build_parser():
@@ -35,24 +35,29 @@ def build_parser():
         default=None,
         help="Directory for dumping results to files",
     )
-    subparsers = parser.add_subparsers(dest="command", help="Analysis subcommands")
+    groups = parser.add_subparsers(dest="group", help="Command groups")
 
-    # db-info
-    subparsers.add_parser("db-info", help="Show database contents: available jobs, time ranges, and metrics")
+    # ---- db group ----
+    db_parser = groups.add_parser("db", help="Database-level commands")
+    db_subs = db_parser.add_subparsers(dest="command", help="Database subcommands")
 
-    # job-info
-    p_info = subparsers.add_parser("job-info", help="Discover job time range and topology")
-    p_info.add_argument("--job", required=True, help="Job ID")
+    db_subs.add_parser("info", help="Show database contents: available jobs, time ranges, and metrics")
+
+    # ---- job group ----
+    job_parser = groups.add_parser("job", help="Job-level analysis commands")
+    job_parser.add_argument("job", metavar="JOBID", help="Job ID")
+    job_subs = job_parser.add_subparsers(dest="command", help="Job subcommands")
+
+    # info
+    p_info = job_subs.add_parser("info", help="Discover job time range and topology")
     p_info.add_argument("--interval", type=float, default=None, help="Sampling interval (seconds) for time refinement")
 
     # metrics
-    p_metrics = subparsers.add_parser("metrics", help="List available metrics for a job")
-    p_metrics.add_argument("--job", required=True, help="Job ID")
+    p_metrics = job_subs.add_parser("metrics", help="List available metrics for a job")
     p_metrics.add_argument("--categorize", action="store_true", help="Group metrics by category")
 
     # stats
-    p_stats = subparsers.add_parser("stats", help="Compute statistics for job metrics")
-    p_stats.add_argument("--job", required=True, help="Job ID")
+    p_stats = job_subs.add_parser("stats", help="Compute statistics for job metrics")
     p_stats.add_argument(
         "--interval",
         type=float,
@@ -72,19 +77,8 @@ def build_parser():
         help="Filter to a specific aggregation level (valid levels depend on category)",
     )
 
-    # data-check
-    p_data_check = subparsers.add_parser("data-check", help="Validate data collection completeness and timing")
-    p_data_check.add_argument("--job", required=True, help="Job ID")
-    p_data_check.add_argument(
-        "--interval",
-        type=float,
-        default=None,
-        help="Sampling interval (seconds) — used for time-range refinement and gap threshold",
-    )
-
     # health
-    p_health = subparsers.add_parser("health", help="Run health checks on a job")
-    p_health.add_argument("--job", required=True, help="Job ID")
+    p_health = job_subs.add_parser("health", help="Run health checks on a job")
     p_health.add_argument(
         "--interval",
         type=float,
@@ -92,9 +86,17 @@ def build_parser():
         help="Sampling interval (seconds) — used for time-range refinement and health step floor",
     )
 
+    # data-check
+    p_data_check = job_subs.add_parser("data-check", help="Validate data collection completeness and timing")
+    p_data_check.add_argument(
+        "--interval",
+        type=float,
+        default=None,
+        help="Sampling interval (seconds) — used for time-range refinement and gap threshold",
+    )
+
     # timeseries
-    p_ts = subparsers.add_parser("timeseries", help="Export raw time series data")
-    p_ts.add_argument("--job", required=True, help="Job ID")
+    p_ts = job_subs.add_parser("timeseries", help="Export raw time series data")
     p_ts.add_argument("--interval", type=float, required=True, help="Sampling interval (seconds)")
     p_ts.add_argument("--metric", required=True, help="Metric name")
     p_ts.add_argument("--node", default=None, help="Filter by node (instance)")
@@ -102,18 +104,16 @@ def build_parser():
     p_ts.add_argument("--output", default=None, help="Write output to file instead of stdout")
 
     # query
-    p_query = subparsers.add_parser("query", help="Execute arbitrary PromQL with job context")
-    p_query.add_argument("--job", required=True, help="Job ID")
+    p_query = job_subs.add_parser("query", help="Execute arbitrary PromQL with job context")
     p_query.add_argument("--interval", type=float, required=True, help="Sampling interval (seconds)")
     p_query.add_argument("--promql", required=True, help="PromQL query ($job and $step are substituted)")
     p_query.add_argument("--step", default=None, help="Query step override (e.g., '5s', '1m')")
     p_query.add_argument("--output", default=None, help="Write output to file instead of stdout")
 
     # iterations
-    p_iter = subparsers.add_parser(
+    p_iter = job_subs.add_parser(
         "iterations", help="Detect iteration boundaries and compute per-iteration statistics"
     )
-    p_iter.add_argument("--job", required=True, help="Job ID")
     p_iter.add_argument(
         "--interval",
         type=float,
@@ -155,8 +155,7 @@ def build_parser():
     )
 
     # counters
-    p_counters = subparsers.add_parser("counters", help="Discover and summarize hardware performance counters")
-    p_counters.add_argument("--job", required=True, help="Job ID")
+    job_subs.add_parser("counters", help="Discover and summarize hardware performance counters")
 
     return parser
 
@@ -165,28 +164,33 @@ def main():
     parser = build_parser()
     args = parser.parse_args()
 
-    if not args.command:
+    if not args.group:
         parser.print_help()
         sys.exit(1)
 
-    analyzer = AnalyzeJob(args.tsdb_url)
+    if not args.command:
+        print(json.dumps({"error": f"No subcommand specified for '{args.group}'. Use --help for available subcommands."}))
+        sys.exit(1)
+
+    analyzer = JobInspector(args.tsdb_url)
 
     handlers = {
-        "db-info": cmd_db_info,
-        "job-info": cmd_job_info,
-        "metrics": cmd_metrics,
-        "stats": cmd_stats,
-        "data-check": cmd_data_check,
-        "health": cmd_health,
-        "timeseries": cmd_timeseries,
-        "query": cmd_query,
-        "iterations": cmd_iterations,
-        "counters": cmd_counters,
+        ("db", "info"): cmd_db_info,
+        ("job", "info"): cmd_job_info,
+        ("job", "metrics"): cmd_metrics,
+        ("job", "stats"): cmd_stats,
+        ("job", "data-check"): cmd_data_check,
+        ("job", "health"): cmd_health,
+        ("job", "timeseries"): cmd_timeseries,
+        ("job", "query"): cmd_query,
+        ("job", "iterations"): cmd_iterations,
+        ("job", "counters"): cmd_counters,
     }
 
-    handler = handlers.get(args.command)
+    key = (args.group, args.command)
+    handler = handlers.get(key)
     if not handler:
-        print(json.dumps({"error": f"Unknown command: {args.command}"}))
+        print(json.dumps({"error": f"Unknown command: {args.group} {args.command}"}))
         sys.exit(1)
 
     try:
@@ -204,7 +208,7 @@ def main():
     # Write to scratch dir if specified
     if args.scratch_dir:
         job = getattr(args, "job", None)
-        filename = f"{args.command}_{job}.json" if job else f"{args.command}.json"
+        filename = f"{args.group}_{args.command}_{job}.json" if job else f"{args.group}_{args.command}.json"
         filepath = _write_scratch(args.scratch_dir, filename, result)
         _append_query_log(args.scratch_dir, analyzer.get_query_stats())
         print(
