+++ omnistat/inspect/constants.py
