+++ omnistat/inspect/output.py
