+++ pyproject.toml
@@ -18,7 +18,7 @@ requires = ["setuptools>=61", "wheel", "setuptools-git-versioning>=2.0,<3", "cma
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-packages = ["omnistat", "omnistat.contrib", "omnistat.inspect"]
+packages = ["omnistat", "omnistat.contrib", "omnistat.inspect", "omnistat.inspect.job"]
 
 [tool.setuptools.dynamic]
 dependencies = { file = ["requirements.txt"] }
